Derive repo name from directory.
Override that value if the repo of this git directory isn't the same
as the directory's base name.

Add secret test to inject an override secret at run time so that no
secret is required at image create time.

Comment out inclusion of additional keys but change permissions for
ssh keys.

Create items in a per instance namespace.